Rename MixpanelService to AnalyticsService
The web-app no longer calls the Mixpanel client in the user's browser,
and instead hits stela's `event` endpoint. While stela currently
forwards those events to Mixpanel, it doesn't really matter what
analytics provider stela is calling out to anymore. Rename this service
to the more generic AnalyticsService.

PER-9642: Update checklist when tasks completed
